backend/requirements.txt
accelerate==1.7.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
apify_fingerprint_datapoints==0.0.2
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
binaryornot==0.4.4
blessed==1.21.0
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
browserforge==1.2.3
cachetools==6.0.0
catalogue==2.0.10
cbor==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
cookiecutter==2.6.0
crawlee==0.6.9
cryptography==45.0.4
cssselect==1.3.0
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
curl_cffi==0.11.1
cymem==2.0.11
datasets==3.6.0
dill==0.3.8
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
editor==1.6.6
espeakng-loader==0.2.4
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
feedparser==6.0.11
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
FlagEmbedding==1.3.4
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
googlenewsdecoder==0.1.7
graphql-core==3.2.6
greenlet==3.2.2
grpcio==1.67.1
h11==0.16.0
h2==4.2.0
h3==4.2.2
hf-xet==1.1.2
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
inquirer==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.10
isodate==0.7.2
itsdangerous==2.2.0
jaro-winkler==2.0.3
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kokoro==0.9.4
langcodes==3.5.0
langdetect==1.0.9
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
lz4==4.4.4
macholib==1.16.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
milvus-lite==2.4.12
misaki==0.9.4
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
nanoid==2.0.0
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numpy==2.2.6
ollama==0.4.8
onnxruntime==1.22.0
openai==1.82.0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
peft==0.15.2
phonemizer-fork==3.3.2
pillow==11.2.1
playwright==1.52.0
preshed==3.0.9
propcache==0.3.1
Protego==0.4.0
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.6.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
pyinstaller==6.14.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pymilvus==2.5.10
pymilvus.model==0.3.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
readability==0.3.2
readability-lxml==0.8.4.1
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
runs==1.2.2
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
segments==2.3.0
selectolax==0.3.29
sentence-transformers==4.1.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
strawberry-graphql==0.270.2
sympy==1.14.0
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
timezonefinder==6.5.9
tinydb==4.8.2
tldextract==5.3.0
tokenizers==0.19.1
torch==2.7.0
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
typer==0.15.4
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
tzdata==2025.2
ujson==5.10.0
unlzw3==0.2.3
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xmod==1.8.1
xxhash==3.5.0
yarl==1.20.0
zlib-state==0.1.9
zstandard==0.23.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85